lib: port read_file over to rest api
GitHub has started to rate limit the usercontent endpoint without a way
to bypass it with authentication this leads to some of our CI jobs
error'ing out with http error 429.

The rest api can also easily obtain small objects and since we
authenticate with a token it is not rate limited.